# swagger.yaml
basePath: /api/v1
consumes:
- application/json
definitions:
  PermissionLevel:
    description: Defines the level of permissions a user has. Certain routes can only
      be used by specific permissions.
    title: PermissionLevel
    type: string
    x-go-package: github.com/MMarsolek/AuctionHouse/model
...
host: localhost
info:
  description: |-
    The purpose of this application is to provide the client with functionality to define users, items for auctions,
    and ways to place bids on those items.
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
  termsOfService: There are no terms of service at the moment. Use at your own risk,
    we take no responsibility.
  title: AuctionHouse API.
  version: 0.1.0
paths:
...
  /api/v1/users:
    post:
      consumes:
      - application/json
      description: This will create a new user and add them to storage. The password
        is hashed before being stored and the text is never logged.
      operationId: postUserRequest
      produces:
      - application/json
      responses:
        "201":
          $ref: '#/responses/noBody'
        "400":
          $ref: '#/responses/errorMessage'
      schemes:
      - http
      summary: Creates a new user that will be specified by the username.
      tags:
      - Users
  /api/v1/users/{username}:
    get:
      description: This will retrieve a user from storage based on the username.
      operationId: getUserRequest
      parameters:
      - description: Username of the user.
        in: path
        name: username
        required: true
        type: string
        x-go-name: Username
      - description: Expected to be "Bearer <auth_token>"
        in: header
        name: Authentication
        type: string
      produces:
      - application/json
      responses:
        "200":
          $ref: '#/responses/getUserResponse'
        "404":
          $ref: '#/responses/noBody'
      schemes:
      - http
      security:
      - api_key: []
      summary: Gets the user specified by the username.
      tags:
      - Users
  /api/v1/users/login:
    post:
      consumes:
      - application/json
      description: This will generate a new authentication token for the user specified
        by the username and password.
      operationId: postLoginRequest
      produces:
      - application/json
      responses:
        "200":
          $ref: '#/responses/noBody'
        "400":
          $ref: '#/responses/errorMessage'
        "404":
          $ref: '#/responses/errorMessage'
      schemes:
      - http
      summary: Retrieves an authentication token for the user.
      tags:
      - Users
  /api/v1/ws:
    get:
      description: |-
        This will upgrade the connection to a websocket. Currently the only command available is to place a bid on an item
        using the command model. See the model defined as WSCommandMessage using WSCommandMessagePlaceBidRequest as the
        payload for how to place a bid using the websocket client.

        All messages are sent back via websocket as the model WSResponseMessage. The Data field inside of the model varies
        based on command. For example, a PlaceBid command defines the Data field as a WSResponseMessagePlaceBidData model.
      operationId: wsRequest
      parameters:
      - description: Expected to be "Bearer <auth_token>"
        in: header
        name: Authentication
        type: string
      produces:
      - application/json
      responses:
        "101":
          $ref: '#/responses/wsConnection'
      schemes:
      - ws
      security:
      - api_key: []
      summary: Establishes a connection via websockets.
      tags:
      - WebSockets
produces:
- application/json
responses:
  errorMessage:
    description: Response when there is an error with the request.
    schema:
      properties:
        message:
          description: The human readable message to the client.
          type: string
          x-go-name: Message
      type: object
...
  getUserResponse:
    description: Contains data about the user and how to identify them.
    schema:
      properties:
        displayName:
          description: The human readable display name of the user.
          type: string
          x-go-name: DisplayName
        permission:
          $ref: '#/definitions/PermissionLevel'
        username:
          description: The username for the user.
          type: string
          x-go-name: Username
      required:
      - username
      - permission
      type: object
  noBody:
    description: Response contains no body.
...
  wsConnection:
    description: Response contains no body.
schemes:
- http
securityDefinitions:
  api_key:
    in: header
    name: KEY
    type: apiKey
swagger: "2.0"